requirements.txt
opencv-python
pillow
image_slicer
numpy
pyautogui
py-sudoku
